+++ services/orchest-webserver/client/src/pipeline-view/hooks/useIsReadOnly.tsx
