+++ services/orchest-webserver/client/src/pipeline-view/hooks/useSocketIO.tsx
@@ -1,8 +1,12 @@
 import React from "react";
 import io from "socket.io-client";
 
+export type SocketIO = Record<"on" | "off" | "emit", any> & {
+  disconnect: () => void;
+};
+
 export const useSocketIO = () => {
-  const [sio, setSio] = React.useState<{ disconnect: () => void }>(null);
+  const [sio, setSio] = React.useState<SocketIO>(null);
   // TODO: only make state.sio defined after successful
   // connect to avoid .emit()'ing to unconnected
   // sio client (emits aren't buffered).
