+++ services/orchest-webserver/client/src/pipeline-view/step-details/StepDetailsLogs.tsx
@@ -1,0 +1,8 @@
+import React from "react";
+import LogViewer, { ILogViewerProps } from "../LogViewer";
+
+export const StepDetailsLogs: React.FC<ILogViewerProps> = (props) => (
+  <div className={"detail-subview"}>
+    <LogViewer {...props} />
+  </div>
+);
